extracted search function to separate file
